+++ src/pages/LoginPage/Forms/ForgotPasswordForm.jsx
@@ -1,23 +1,22 @@
 import React from "react";
-import { PasswordResetRequest } from '../../db/passwordResetRequest';
-import { NotificationMessage } from "./NotifcationMessage";
-import { ErrorMessage } from './ErrorMessage';
-import { CheckInputIsntEmpty } from "../../validation/inputValidation";
-import { IsEmailValid } from "../../validation/emailValidation";
-import { Button } from './../../components/Button/Button';
+import { PasswordResetRequest } from '../../../db/passwordResetRequest';
+import { NotificationMessage } from "../NotifcationMessage";
+import { ErrorMessage } from '../ErrorMessage';
+import { CheckInputIsntEmpty } from "../../../validation/inputValidation";
+import { IsEmailValid } from "../../../validation/emailValidation";
+import { Button } from '../../../components/Button/Button';
+import { TemplateForm } from "./TemplateForm";
 
-import "../../App.css";
-import "../LoginPage/LoginPage.css";
+import "../../../App.css";
+import "../LoginPage.css";
+
 
 //This is the form we enter our email into
 
-export const ForgotPasswordForm = ({setResetPassword}) => {
-
-
+const Form = ({setResetPassword}) => {
     return (
-        <div id="forgotPassword-entry" className="center-content" style={{margin: "50px auto 0 auto"}}>
-            <div id="forgotPassword-content" className="login-form forgotPassword-form">
-                <h2 className="loginInput-title">Password Reset Request</h2>
+        <>
+            <h2 className="loginInput-title">Password Reset Request</h2>
                 <h3 className="forgotPassword-subtitle">Email</h3>
                 <input id="forgotPasswordInput" className="login-input" type="text"></input>
                 <Button text="Send Email" func={() => {
@@ -39,10 +38,15 @@ export const ForgotPasswordForm = ({setResetPassword}) => {
                         })
                     }   
                 }}/>
-                <Button text="Back" func={() => {
-                    setResetPassword(false);
-                }}/>
-            </div>
-        </div>
+            <Button text="Back" func={() => {
+                setResetPassword(false);
+            }}/>
+        </>
+    )
+}
+
+export const ForgotPasswordForm = ({setResetPassword}) => {
+    return (
+        <TemplateForm FormContent={<Form setResetPassword={setResetPassword}/>}/>
     )
 }
